producer/requirements.txt
kafka-python==2.0.2
faker
numpy
python-dateutil
pyyaml
